Revert "feat(): refactoring of models structure and code implementations for users-organizations flow"
