+++ bigfastapi/core/helpers.py
@@ -2,7 +2,8 @@ import requests
 import sqlalchemy.orm as _orm
 from decouple import config
 
-from bigfastapi.models.organization_models import Organization, OrganizationUser
+from bigfastapi.models import organization_user_model
+from bigfastapi.models.organization_models import Organization
 
 
 class Helpers:
@@ -14,7 +15,7 @@ class Helpers:
                 .first()
         )
 
-        store_user = db.query(OrganizationUser).filter_by(store_id=organization_id).filter_by(
+        store_user = db.query(organization_user_model.organizationUser).filter_by(store_id=organization_id).filter_by(
             user_id=user_id).first()
         if store_user == None and organization == None:
             return False
@@ -22,7 +23,7 @@ class Helpers:
 
     # Sends a notification to slack.
     # NOTE: DO NOT CALL THIS METHOD IN THE SAME THREAD AS YOUR REQUEST. USE A BACKGROUND TASK
-    @staticmethod   
+    @staticmethod
     def slack_notification(url: str, text: str, verify: bool = True):
         requests.post(url=config(url), json={"text": text}, headers={"Content-Type": "application/json"},
                     verify=verify)
