+++ bigfastapi/credit.py
@@ -16,7 +16,7 @@ from starlette.responses import RedirectResponse
 from bigfastapi.db.database import get_db
 from .auth_api import is_authenticated
 from .core.helpers import Helpers
-from .models import credit_wallet_models as model, organization_models, credit_wallet_conversion_models, wallet_models, \
+from .models import credit_wallet_models as model, organization_models, credit_wallet_conversion_models, organization_user_model, wallet_models, \
     wallet_transaction_models, credit_wallet_history_models
 from .schemas import credit_wallet_schemas as schema, credit_wallet_conversion_schemas
 from .schemas import users_schemas
