+++ bigfastapi/organization.py
@@ -1,32 +1,28 @@
 import datetime as _dt
 import os
-from typing import Optional
 from uuid import uuid4
 
 import fastapi as _fastapi
-from fastapi.responses import JSONResponse
 import sqlalchemy.orm as _orm
 from decouple import config
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException
 from fastapi import BackgroundTasks
 from fastapi import UploadFile, File
 from fastapi.responses import FileResponse
 from sqlalchemy import and_
 
 from bigfastapi.db.database import get_db
-from bigfastapi.email import send_email
 from bigfastapi.models import organization_models
 from bigfastapi.models.menu_model import addDefaultMenuList, getOrgMenu
+from bigfastapi.schemas import roles_schemas
 from .auth_api import is_authenticated
 from .core.helpers import Helpers
 from .files import upload_image
-from .models import credit_wallet_models as credit_wallet_models
-from .schemas import organization_schemas as _schemas
-from .schemas.organization_schemas import _OrganizationBase, AddRole, OrganizationUserBase
-from .models.organization_models import OrganizationInvite, OrganizationUser, Role
+from .models import credit_wallet_models as credit_wallet_models, organization_invite_model, organization_user_model
 from .models import organization_models as _models
-from .models import user_models, schedule_models
+from .models import user_models, role_models, schedule_models
 from .models import wallet_models as wallet_models
+from .schemas import organization_schemas as _schemas
 from .schemas import users_schemas
 from .utils.utils import paginate_data, row_to_dict
 
@@ -236,7 +232,7 @@ async def get_organization(
     return {"data": {"organization": organization, "menu": menu}}
 
 
-@app.get("/organizations/{organization_id}/users", status_code=200, response_model=_schemas.OrganizationUsersResponse)
+@app.get("/organizations/{organization_id}/users", status_code=200)
 async def get_organization_users(
         organization_id: str,
         db: _orm.Session = _fastapi.Depends(get_db)
@@ -251,10 +247,11 @@ async def get_organization_users(
 
         returnBody--> list of all users in the queried organization
     """
-    invited_list = db.query(OrganizationUser).filter(
+    # query the store_users table with the organization_id
+    invited_list = db.query(organization_user_model.organizationUser).filter(
         and_(
-            OrganizationUser.organization_id == organization_id,
-            OrganizationUser.is_deleted == False
+            organization_user_model.organizationUser.store_id == organization_id,
+            organization_user_model.organizationUser.is_deleted == False
         )
     ).all()
 
@@ -269,10 +266,10 @@ async def get_organization_users(
     if organization is None:
         raise _fastapi.HTTPException(
             status_code=404, detail="Organization does not exist")
-    organization_owner_id = organization.creator
-    organization_owner = (
+    store_owner_id = organization.creator
+    store_owner = (
         db.query(user_models.User)
-            .filter(user_models.User.id == organization_owner_id)
+            .filter(user_models.User.id == store_owner_id)
             .first())
 
     invited_users = []
@@ -283,18 +280,16 @@ async def get_organization_users(
                     user_models.User.id == invited["user_id"],
                     user_models.User.is_deleted == False
                 )).first()
-            role = db.query(Role).filter(
-                Role.id == invited["role_id"]).first()
+            role = db.query(role_models.Role).filter(
+                role_models.Role.id == invited["role_id"]).first()
             if (user.id == invited["user_id"]):
-                invited["first_name"] = user.first_name
-                invited["last_name"] = user.last_name
                 invited["email"] = user.email
                 invited["role"] = role.role_name
 
             invited_users.append(invited)
 
     users = {
-        "user": organization_owner,
+        "user": store_owner,
         "invited": invited_users
     }
     return users
@@ -313,43 +308,35 @@ def delete_organization_user(
             param--> user_id: This is the unique id of the user to be removed from the organization
 
     returnDesc--> On sucessful request, it returns message,
-        returnBody--> User with email {email} successfully removed from the organization
+        returnBody--> User with email {email} successfully removed from the store
     """
-    try:
+    # fetch the organization user from the user table
+    user = db.query(user_models.User).filter(
+        user_models.User.id == user_id).first()
 
-        # fetch the organization user from the user table
-        user = db.query(user_models.User).filter(
-            user_models.User.id == user_id).first()
+    if user is not None:
+        # fetch the store user from the store user table.
+        store_user = (
+            db.query(organization_user_model.organizationUser)
+                .filter(and_(
+                organization_user_model.organizationUser.user_id == user_id,
+                organization_user_model.organizationUser.store_id == organization_id
+            ))
+                .first()
+        )
 
-        if user is not None:
-            organization_user = (
-                db.query(OrganizationUser)
-                    .filter(and_(
-                    OrganizationUser.user_id == user_id,
-                    OrganizationUser.organization_id == organization_id
-                ))
-                    .first()
-            )
+        store_user.is_deleted = True
+        user.is_deleted = True
+        db.add(store_user)
+        db.commit()
+        db.refresh(store_user)
+        db.add(user)
+        db.commit()
+        db.refresh(user)
 
-            organization_user.is_deleted = True
-            user.is_deleted = True
-            db.add(organization_user)
-            db.commit()
-            db.refresh(organization_user)
-            db.add(user)
-            db.commit()
-            db.refresh(user)
+        return {"message": f"User with email {user.email} successfully removed from the store"}
 
-            return {"message": f"User with email {user.email} successfully removed from the organization"}
-
-        return {"message": "User does not exist"}
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
-
+    return {"message": "User does not exist"}
 
 
 @app.get("/organizations/{organization_id}/roles")
@@ -363,15 +350,15 @@ def get_roles(organization_id: str, db: _orm.Session = _fastapi.Depends(get_db))
 
         returnBody--> list of all available roles in the queried organization
     """
-    roles = db.query(Role).filter(
-        Role.organization_id == organization_id)
-    roles = list(map(_schemas.Role.from_orm, roles))
+    roles = db.query(role_models.Role).filter(
+        role_models.Role.organization_id == organization_id)
+    roles = list(map(roles_schemas.Role.from_orm, roles))
 
     return roles
 
 
 @app.post("/organizations/{organization_id}/roles")
-def add_role(payload: AddRole,
+def add_role(payload: roles_schemas.AddRole,
              organization_id: str,
              db: _orm.Session = _fastapi.Depends(get_db)
              ):
@@ -386,301 +373,33 @@ def add_role(payload: AddRole,
     returnDesc--> On sucessful request, it returns:
         returnBody--> details of the newly created organization role.
     """
-    try:
-
-        roles = db.query(Role).filter(
-            Role.organization_id == organization_id
-        ).all()
-        if len(roles) < 1:
-            existing_role = (
-                db.query(Role)
-                    .filter(Role.role_name == payload.role_name.lower())
-                    .first()
+    roles = db.query(role_models.Role).filter(
+        role_models.Role.organization_id == organization_id
+    ).all()
+    if len(roles) < 1:
+        existing_role = (
+            db.query(role_models.Role)
+                .filter(role_models.Role.role_name == payload.role_name.lower())
+                .first()
+        )
+        if existing_role is None:
+            role = role_models.Role(
+                id=uuid4().hex,
+                organization_id=organization_id.strip(),
+                role_name=payload.role_name.lower()
             )
-            if existing_role is None:
-                role = Role(
-                    id=uuid4().hex,
-                    organization_id=organization_id.strip(),
-                    role_name=payload.role_name.lower()
-                )
 
-                db.add(role)
-                db.commit()
-                db.refresh(role)
+            db.add(role)
+            db.commit()
+            db.refresh(role)
 
-                return role
-            return {"message": "role already exist"}
-        return
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
+            return role
+
+        return {"message": "role already exist"}
+    return
 
 
-
-@app.put('/organizations/accept-invite/{invite_code}', response_model=_schemas.AcceptInviteResponse)
-def accept_invite(
-        payload: _schemas.OrganizationUser,
-        invite_code: str,
-        db: _orm.Session = _fastapi.Depends(get_db)):
-    """intro-->This endpoint allows for a user to accept an invite. To use this endpoint you need to make a put request to the /users/accept-invite/{token} where token is a unique value recieved by the user on invite. It also takes a specified body of request
-
-    paramDesc-->On put request this enpoint takes the query parameter "token" 
-        param-->token: This is a unique token recieved by the user on invite
-        reqBody-->user_email: This is the email address of the user 
-        reqBody-->user_id: This is the unique user id
-        reqBody-->user_role: This specifies the role of the user in the organization  
-        reqBody-->is_accepted: This is the the acceptance state of the invite  
-        reqBody-->is_revoked: This is the revoke state of the user  
-        reqBody-->is_deleted: This specifies if the invite is deleted/expired  
-        reqBody-->organization_id: This is a unique id of the registered organization
-
-    returnDesc--> On sucessful request, it returns:
-
-        returnBody--> An object with a key `invited` containing the new organization user data, and `organization` containing information about the organization
-         the user is invited to.
-    """
-    try: 
-
-        existing_invite = db.query(
-            OrganizationInvite).filter(
-                and_(
-                    OrganizationInvite.invite_code == invite_code,
-                    OrganizationInvite.is_deleted == False,
-                    OrganizationInvite.is_revoked == False
-                )).first()
-
-        if existing_invite is None:
-            return JSONResponse({
-                "message": "Invite not found. If you think this is shouldn't be, try again or ask the inviter to invite you again."
-            }, status_code=404)
-
-        organization = db.query(organization_models.Organization).filter(
-            organization_models.Organization.id == existing_invite.organization_id).first()
-
-        organization_user = OrganizationUser(
-            id=uuid4().hex,
-            organization_id=existing_invite.organization_id,
-            user_id=payload.user_id,
-            role_id=existing_invite.role_id
-        )
-        db.add(organization_user)
-        db.commit()
-        db.refresh(organization_user)
-
-        existing_invite.is_deleted = True
-        existing_invite.is_accepted = True
-        db.commit()
-        db.refresh(existing_invite)
-
-        return { "invited": OrganizationUserBase.from_orm(organization_user), "organization": _OrganizationBase.from_orm(organization) }
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex))
-
-
-
-@app.post("/organizations/{organization_id}/invite-user/", status_code=201, response_model=_schemas.InviteResponse)
-async def invite_user(
-    payload: _schemas.UserInvite,
-    background_tasks: BackgroundTasks,
-    template: Optional[str] = "invite_email.html",
-    user: str = _fastapi.Depends(is_authenticated),
-    db: _orm.Session = _fastapi.Depends(get_db)
-):
-    """intro--> This endpoint is used to trigger a user invite. To use this endpoint you need to make a post request to the /users/invite/ endpoint with the specified body of request 
-
-        reqBody--> user_email: This is the email address of the user to be invited.
-        reqBody--> user_id: This is the unique user id of the logged in user
-        reqBody--> user_role: This specifies the role of the user to be invited in the organization   
-        reqBody--> organization: This specifies the information of the registered organization
-        reqBody--> app_url: This is the url to be navigated to on invite accept, usually the url of the application.
-        reqBody--> email_details: This is the key content of the invite email to be sent.
-
-    'returnDesc'--> On sucessful request, it returns:
-
-        returnBody-->  An object with a key `message`.
-    """
-    try:
-        invite_token = uuid4().hex
-        invite_url = f"{payload.app_url}/accept-invite?invite_token={invite_token}"
-        payload.email_details.link = invite_url
-        email_info = payload.email_details
-
-        role = (
-            db.query(Role)
-            .filter(Role.role_name == payload.user_role.lower())
-            .first()
-        )
-
-        # make sure you can't send invite to yourself
-        if (user.email != payload.user_email):
-            existing_invite = (
-                db.query(OrganizationInvite)
-                .filter(
-                    and_(
-                        OrganizationInvite.user_email == payload.user_email,
-                        OrganizationInvite.is_deleted == False
-                    )).first())
-            if existing_invite is None:
-
-                send_email(email_details=email_info,
-                        background_tasks=background_tasks, template=template, db=db)
-                invite = OrganizationInvite(
-                    id=uuid4().hex,
-                    organization_id=payload.organization.get("id"),
-                    user_id=payload.user_id,
-                    user_email=payload.user_email,
-                    role_id=role.id,
-                    invite_code=invite_token
-                )
-                db.add(invite)
-                db.commit()
-                db.refresh(invite)
-
-                return {"message": "Organization invite email will be sent in the background."}
-            return {"message": "invite already sent"}
-        return {"message": "Enter an email you're not logged in with."}
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
-
-
-@app.get('/organizations/get-invite/{invite_code}', response_model=_schemas.SingleInviteResponse)
-async def get_single_invite(
-    invite_code: str,
-    db: _orm.Session = _fastapi.Depends(get_db),
-):
-    """intro--> This endpoint is used to get an invite link for a single user. To use this endpoint you need to make a get request to the /users/invite/{invite_code} endpoint
-
-    paramDesc--> On get request, the url takes an invite code
-        param--> invite_code: This is a unique code needed to get an invite link
-
-
-    returnDesc--> On sucessful request, it returns
-        returnBody--> An object with a key `invite` containing the invite data and a key `user` containing an empty string `''`
-        indicating the user is not a member of another organization in the application or `exists` indicating that the invited user is a member of another organization in the application.
-    """
-    try:
-
-        # user invite code to query the invite table
-        existing_invite = db.query(
-            OrganizationInvite).filter(
-                and_(
-                    OrganizationInvite.invite_code == invite_code,
-                    OrganizationInvite.is_deleted == False,
-                    OrganizationInvite.is_revoked == False
-                )).first()
-        if (existing_invite):
-            existing_user = db.query(user_models.User).filter(
-                user_models.User.email == existing_invite.user_email).first()
-
-            organization = db.query(organization_models.Organization).filter(
-                organization_models.Organization.id == existing_invite.organization_id).first()
-
-            setattr(existing_invite, 'organization', organization)
-            user_exists = ''
-            if(existing_user is not None):
-                user_exists = 'exists'
-            if not existing_invite:
-                return JSONResponse({
-                    "message": "Invite not found! Try again or ask the inviter to invite you again."
-                }, status_code=404)
-
-            return {"invite": existing_invite, "user": user_exists}
-        return JSONResponse({
-            "message": "Invalid invite code"
-        }, status_code=404)
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
-
-
-@app.put("/organizations/decline-invite/{invite_code}", response_model=_schemas.DeclinedInviteResponse)
-def decline_invite(
-    invite_code: str,
-    db: _orm.Session = _fastapi.Depends(get_db)):
-    """intro--> This endpoint is used to decline an invite. To use this endpoint you need to make a put request to the /users/invite/{invite_code}/decline endpoint
-
-    paramDesc--> On put request, the url takes an invite code
-        param-->invite_code: This is a unique code linked to invite
-
-
-    returnDesc--> On sucessful request, it returns message:
-
-        returnBody--> an object contain the invite data with the `is_deleted` field set to True
-    """
-    try:
-
-        declined_invite = (
-            db.query(OrganizationInvite)
-            .filter(OrganizationInvite.invite_code == invite_code)
-            .first()
-        )
-
-        declined_invite.is_deleted = True
-        db.add(declined_invite)
-        db.commit()
-        db.refresh(declined_invite)
-
-        return declined_invite
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
-
-
-@app.delete("/organizations/{organization_id}/revoke-invite/{invite_code}", response_model=_schemas.RevokedInviteResponse)
-def revoke_invite(
-    organization_id: str,
-    invite_code: str,
-    db: _orm.Session = _fastapi.Depends(get_db)
-):
-    """intro-->This endpoint is used to revoke the invitation of a previously invited user. To use this endpoint you need to make a delete request to the /users/revoke-invite/{invite_code} endpoint
-
-    paramDesc-->On delete request, the url takes an invite code
-        param-->invite_code: This is a unique code linked to invite
-
-
-    returnDesc--> On successful request, it returns message,
-        returnBody--> an object contain the invite data with the `is_deleted` and `is_revoked` field set to True
-    """
-    try:
-
-        revoked_invite = (
-            db.query(OrganizationInvite)
-            .filter(and_(OrganizationInvite.organization_id == organization_id, OrganizationInvite.invite_code == invite_code))
-            .first()
-        )
-
-        revoked_invite.is_revoked = True
-        revoked_invite.is_deleted = True
-        db.add(revoked_invite)
-        db.commit()
-        db.refresh(revoked_invite)
-
-        return revoked_invite
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
-
-
-@app.get("/organizations/{organization_id}/invites", status_code=200, response_model=_schemas.AllInvites)
+@app.get("/organizations/invites/{organization_id}")
 def get_pending_invites(
         organization_id: str,
         db: _orm.Session = _fastapi.Depends(get_db)
@@ -695,25 +414,18 @@ def get_pending_invites(
 
         returnBody--> all pending invites in the queried organization
     """
-    try:
-        pending_invites = (
-            db.query(OrganizationInvite)
-                .filter(
-                and_(OrganizationInvite.organization_id == organization_id,
-                    OrganizationInvite.is_deleted == False,
-                    OrganizationInvite.is_accepted == False,
-                    OrganizationInvite.is_revoked == False
-                    ))
-                .all()
-        )
+    pending_invites = (
+        db.query(organization_invite_model.organizationInvite)
+            .filter(
+            and_(organization_invite_model.organizationInvite.store_id == organization_id,
+                 organization_invite_model.organizationInvite.is_deleted == False,
+                 organization_invite_model.organizationInvite.is_accepted == False,
+                 organization_invite_model.organizationInvite.is_revoked == False
+                 ))
+            .all()
+    )
 
-        return { "data": pending_invites }
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(ex)) 
-    
+    return pending_invites
 
 
 @app.put("/organizations/{organization_id}")
@@ -858,7 +570,7 @@ def create_organization(user: users_schemas.User, db: _orm.Session, organization
 def runWalletCreation(newOrganization: organization_models.Organization, db: _orm.Session):
     roles = ["Assistant", "Admin", "Owner"]
     for role in roles:
-        new_role = Role(
+        new_role = role_models.Role(
             id=uuid4().hex,
             organization_id=newOrganization.id.strip(),
             role_name=role.lower()
@@ -877,8 +589,8 @@ def get_organizations(user: users_schemas.User, db: _orm.Session):
         creator=user.id).all()
 
     invited_orgs_rep = (
-        db.query(OrganizationUser)
-            .filter(OrganizationUser.user_id == user.id)
+        db.query(organization_user_model.organizationUser)
+            .filter(organization_user_model.organizationUser.user_id == user.id)
             .all()
     )
 
@@ -895,13 +607,13 @@ def get_organizations(user: users_schemas.User, db: _orm.Session):
 
         return organizationCollection
 
-    organization_id_list = list(map(lambda x: x.organization_id, invited_orgs_rep))
+    store_id_list = list(map(lambda x: x.store_id, invited_orgs_rep))
 
     org = []
-    for org_id in organization_id_list:
+    for store_id in store_id_list:
         org = org + \
               db.query(_models.Organization).filter(
-                  _models.Organization.id == org_id).all()
+                  _models.Organization.id == store_id).all()
 
     org_coll = native_orgs + org
     organizationCollection = []
