+++ bigfastapi/users.py
@@ -1,18 +1,26 @@
+from operator import inv
+from re import L
+from typing import Optional
+import uuid
+from uuid import uuid4
 import fastapi as fastapi
+from fastapi.responses import JSONResponse
 from sqlalchemy import and_
 import os
 
 import passlib.hash as _hash
-from bigfastapi.models import user_models, auth_models
-from fastapi import APIRouter, HTTPException, UploadFile, File, status
+from bigfastapi.models import organization_models, user_models, auth_models
+from fastapi import APIRouter, HTTPException, UploadFile, File, BackgroundTasks, status
 import sqlalchemy.orm as orm
 from bigfastapi.db.database import get_db
 from .schemas import users_schemas as _schemas
-from .schemas.organization_schemas import OrganizationUserBase, RoleUpdate, UpdateRoleResponse
+from .schemas import organization_invite_schemas as _invite_schemas
+from .schemas.organization_user_schemas import RoleUpdate, UpdateRoleResponse, organizationUserBase
+from .schemas.organization_schemas import _OrganizationBase
 from .auth_api import is_authenticated, send_code_password_reset_email,  resend_token_verification_mail, verify_user_token, password_change_token
 from .files import deleteFile, isFileExist, upload_image
 from .email import send_email
-from .models.organization_models import OrganizationUser, Role
+from .models import organization_invite_model, organization_user_model, role_models
 
 
 app = APIRouter(tags=["User"])
@@ -150,6 +158,250 @@ async def updatePassword(
     return {"data":  dbResponse}
 
 
+@app.put('/users/accept-invite/{token}', response_model=_invite_schemas.AcceptInviteResponse)
+def accept_invite(
+        payload: _invite_schemas.organizationUser,
+        token: str,
+        db: orm.Session = fastapi.Depends(get_db)):
+    """intro-->This endpoint allows for a user to accept an invite. To use this endpoint you need to make a put request to the /users/accept-invite/{token} where token is a unique value recieved by the user on invite. It also takes a specified body of request
+
+    paramDesc-->On put request this enpoint takes the query parameter "token" 
+        param-->token: This is a unique token recieved by the user on invite
+        reqBody-->user_email: This is the email address of the user 
+        reqBody-->user_id: This is the unique user id
+        reqBody-->user_role: This specifies the role of the user in the organization  
+        reqBody-->is_accepted: This is the the acceptance state of the invite  
+        reqBody-->is_revoked: This is the revoke state of the user  
+        reqBody-->is_deleted: This specifies if the invite is deleted/expired  
+        reqBody-->organization_id: This is a unique id of the registered organization
+
+    returnDesc--> On sucessful request, it returns:
+
+        returnBody--> An object with a key `invited` containing the new store user data, and `store` containing information about the store
+         the user is invited to.
+    """
+
+    existing_invite = db.query(
+        organization_invite_model.organizationInvite).filter(
+            and_(
+                organization_invite_model.organizationInvite.invite_code == token,
+                organization_invite_model.organizationInvite.is_deleted == False,
+                organization_invite_model.organizationInvite.is_revoked == False
+            )).first()
+
+    if existing_invite is None:
+        return JSONResponse({
+            "message": "Invite not found! Try again or ask the inviter to invite you again."
+        }, status_code=404)
+
+    existing_user = db.query(user_models.User).filter(
+        user_models.User.email == existing_invite.user_email).first()
+
+    if existing_user is None:
+        return JSONResponse({
+            "message": "You must log in first"
+        }, status_code=403)
+
+    # check if the invite token exists in the db.
+    invite = db.query(organization_invite_model.organizationInvite).filter(
+        organization_invite_model.organizationInvite.invite_code == token).first()
+    if invite is None:
+        return JSONResponse({
+            "message": "Invite not found!"
+        }, status_code=status.HTTP_404_NOT_FOUND)
+
+    store = db.query(organization_models.Organization).filter(
+        organization_models.Organization.id == invite.store_id).first()
+
+    # TO-DO
+    # check if the store user exist and update before creating store user
+
+    store_user = organization_user_model.organizationUser(
+        id=uuid4().hex,
+        store_id=payload.organization_id,
+        user_id=payload.user_id,
+        role_id=invite.role_id
+    )
+    db.add(store_user)
+    db.commit()
+    db.refresh(store_user)
+
+    invite.is_deleted = True
+    invite.is_accepted = True
+    db.add(invite)
+    db.commit()
+    db.refresh(invite)
+
+    return {"invited": organizationUserBase.from_orm(store_user), "store": _OrganizationBase.from_orm(store)}
+
+
+@app.post("/users/invite/", status_code=201, response_model=_invite_schemas.InviteResponse)
+async def invite_user(
+    payload: _invite_schemas.UserInvite,
+    background_tasks: BackgroundTasks,
+    template: Optional[str] = "invite_email.html",
+    user: str = fastapi.Depends(is_authenticated),
+    db: orm.Session = fastapi.Depends(get_db)
+):
+    """intro--> This endpoint is used to trigger a user invite. To use this endpoint you need to make a post request to the /users/invite/ endpoint with the specified body of request 
+
+        reqBody--> user_email: This is the email address of the user to be invited.
+        reqBody--> user_id: This is the unique user id of the logged in user
+        reqBody--> user_role: This specifies the role of the user to be invited in the organization   
+        reqBody--> store: This specifies the information of the registered organization
+        reqBody--> app_url: This is the url to be navigated to on invite accept, usually the url of the application.
+        reqBody--> email_details: This is the key content of the invite email to be sent.
+
+    'returnDesc'--> On sucessful request, it returns:
+
+        returnBody-->  An object with a key `message`.
+    """
+
+    invite_token = uuid4().hex
+    invite_url = f"{payload.app_url}/accept-invite?code={invite_token}"
+    payload.email_details.link = invite_url
+    email_info = payload.email_details
+
+    role = (
+        db.query(role_models.Role)
+        .filter(role_models.Role.role_name == payload.user_role.lower())
+        .first()
+    )
+
+    # make sure you can't send invite to yourself
+    if (user.email != payload.user_email):
+        # check if user_email already exists
+        existing_invite = (
+            db.query(organization_invite_model.organizationInvite)
+            .filter(
+                and_(
+                    organization_invite_model.organizationInvite.user_email == payload.user_email,
+                    organization_invite_model.organizationInvite.is_deleted == False
+                )).first())
+        if existing_invite is None:
+
+            # send invite email to user
+            send_email(email_details=email_info,
+                       background_tasks=background_tasks, template=template, db=db)
+            invite = organization_invite_model.organizationInvite(
+                id=uuid4().hex,
+                store_id=payload.store.get("id"),
+                user_id=payload.user_id,
+                user_email=payload.user_email,
+                role_id=role.id,
+                invite_code=invite_token
+            )
+            db.add(invite)
+            db.commit()
+            db.refresh(invite)
+
+            return {"message": "Store invite email will be sent in the background."}
+        return {"message": "invite already sent"}
+    return {"message": "Enter an email you're not logged in with."}
+
+
+@app.get('/users/invite/{invite_code}', response_model=_invite_schemas.SingleInviteResponse)
+async def get_single_invite(
+    invite_code: str,
+    db: orm.Session = fastapi.Depends(get_db),
+):
+    """intro--> This endpoint is used to get an invite link for a single user. To use this endpoint you need to make a get request to the /users/invite/{invite_code} endpoint
+
+    paramDesc--> On get request, the url takes an invite code
+        param--> invite_code: This is a unique code needed to get an invite link
+
+
+    returnDesc--> On sucessful request, it returns
+        returnBody--> An object with a key `invite` containing the invite data and a key `user` containing an empty string `''`
+        indicating the user is not a member of another organization in the application or `exists` indicating that the invited user is a member of another organization in the application.
+    """
+    # user invite code to query the invite table
+    existing_invite = db.query(
+        organization_invite_model.organizationInvite).filter(
+            and_(
+                organization_invite_model.organizationInvite.invite_code == invite_code,
+                organization_invite_model.organizationInvite.is_deleted == False,
+                organization_invite_model.organizationInvite.is_revoked == False
+            )).first()
+    if (existing_invite):
+        existing_user = db.query(user_models.User).filter(
+            user_models.User.email == existing_invite.user_email).first()
+
+        store = db.query(organization_models.Organization).filter(
+            organization_models.Organization.id == existing_invite.store_id).first()
+
+        # existing_invite.__setattr__('store', store)
+        setattr(existing_invite, 'store', store)
+        user_exists = ''
+        if(existing_user is not None):
+            user_exists = 'exists'
+        if not existing_invite:
+            return JSONResponse({
+                "message": "Invite not found! Try again or ask the inviter to invite you again."
+            }, status_code=404)
+
+        return {"invite": existing_invite, "user": user_exists}
+    return JSONResponse({
+        "message": "Invalid invite code"
+    }, status_code=400)
+
+
+@app.put("/users/invite/{invite_code}/decline", response_model=_invite_schemas.DeclinedInviteResponse)
+def decline_invite(invite_code: str, db: orm.Session = fastapi.Depends(get_db)):
+    """intro--> This endpoint is used to decline an invite. To use this endpoint you need to make a put request to the /users/invite/{invite_code}/decline endpoint
+
+    paramDesc--> On put request, the url takes an invite code
+        param-->invite_code: This is a unique code linked to invite
+
+
+    returnDesc--> On sucessful request, it returns message:
+
+        returnBody--> an object contain the invite data with the `is_deleted` field set to True
+    """
+
+    declined_invite = (
+        db.query(organization_invite_model.organizationInvite)
+        .filter(organization_invite_model.organizationInvite.invite_code == invite_code)
+        .first()
+    )
+
+    declined_invite.is_deleted = True
+    db.add(declined_invite)
+    db.commit()
+    db.refresh(declined_invite)
+
+    return declined_invite
+
+
+@app.delete("/users/revoke-invite/{invite_code}", response_model=_invite_schemas.RevokedInviteResponse)
+def revoke_invite(
+    invite_code: str,
+    db: orm.Session = fastapi.Depends(get_db)
+):
+    """intro-->This endpoint is used to revoke the invitation of a previously invited user. To use this endpoint you need to make a delete request to the /users/revoke-invite/{invite_code} endpoint
+
+    paramDesc-->On delete request, the url takes an invite code
+        param-->invite_code: This is a unique code linked to invite
+
+
+    returnDesc--> On successful request, it returns message,
+        returnBody--> an object contain the invite data with the `is_deleted` and `is_revoked` field set to True
+    """
+    revoked_invite = (
+        db.query(organization_invite_model.organizationInvite)
+        .filter(organization_invite_model.organizationInvite.invite_code == invite_code)
+        .first()
+    )
+
+    revoked_invite.is_revoked = True
+    revoked_invite.is_deleted = True
+    db.add(revoked_invite)
+    db.commit()
+    db.refresh(revoked_invite)
+
+    return revoked_invite
+
+
 @app.patch("/users/{user_id}/change", response_model=UpdateRoleResponse)
 def update_user_role(
     payload: RoleUpdate,
@@ -166,44 +418,37 @@ def update_user_role(
         returnBody--> An object with a key `message` with the value - "User role successfully updated", 
             and `data` containing the updated store user data.
     """
-    try:
 
-        existing_user = (
-            db.query(user_models.User)
-            .filter(
-                user_models.User.email == payload.email
-            )
+    existing_user = (
+        db.query(user_models.User)
+        .filter(
+            user_models.User.email == payload.email
+        )
+        .first()
+    )
+
+    if existing_user is not None:
+        existing_store_user = (
+            db.query(organization_user_model.organizationUser)
+            .filter(organization_user_model.organizationUser.user_id == existing_user.id)
             .first()
         )
 
-        if existing_user is not None:
-            existing_store_user = (
-                db.query(OrganizationUser)
-                .filter(OrganizationUser.user_id == existing_user.id)
-                .first()
-            )
+        role = (
+            db.query(role_models.Role)
+            .filter(role_models.Role.role_name == payload.role.lower())
+            .first()
+        )
+        existing_store_user.role_id = role.id
+        db.add(existing_store_user)
+        db.commit()
+        db.refresh(existing_store_user)
 
-            role = (
-                db.query(Role)
-                .filter(Role.role_name == payload.role.lower())
-                .first()
-            )
-            existing_store_user.role_id = role.id
-            db.add(existing_store_user)
-            db.commit()
-            db.refresh(existing_store_user)
-
-            return {
-                "message": "User role updated successfully",
-                "data": OrganizationUserBase.from_orm(existing_store_user)
-            }
-        return {"message": "User does not exist"}
-
-    except Exception as ex:
-        if type(ex) == HTTPException:
-            raise ex
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
-            , detail=str(ex)) 
+        return {
+            "message": "User role successfully updated",
+            "data": existing_store_user
+        }
+    return {"message": "User does not exist"}
 
 # ////////////////////////////////////////////////////CODE //////////////////////////////////////////////////////////////
