+++ bigfastapi/utils/settings.py
@@ -28,9 +28,12 @@ LANDING_PAGE_JS = pkg_resources.resource_filename("bigfastapi","/templates/")
 # PASSWORD_RESET_TEMPLATE="email/password_reset.html"
 
 # If the templates folder is not valid, then we use the bigfastapi templates folder
-if os.path.exists(TEMPLATE_FOLDER) == False:
+if os.path.exists(TEMPLATE_FOLDER) != False:
     TEMPLATE_FOLDER = pkg_resources.resource_filename(
-        'bigfastapi', 'templates/')
+        'bigfastapi', '/templates/email/')
+else:
+    TEMPLATE_FOLDER = pkg_resources.resource_filename(
+        'bigfastapi', '/templates/email/')
 
 # EMAIL_VERIFICATION_TEMPLATE = os.path.join(TEMPLATE_FOLDER, "email", "welcome_email.html")
 # PASSWORD_RESET_TEMPLATE = os.path.join(TEMPLATE_FOLDER,"email", "password_reset.html")
